replace Dto to CloudServer.Common
