+++ CloudServer.Common/Constants/DefaultClaims.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CloudServer.Data.Models
+{
+    public class DefaultClaims
+    {
+        public const string Address = "address";
+        public const string PhoneNumber = "phone_number";
+        public const string Email = "email";
+        public const string Username = "username";
+        public const string UserId = "user_id";
+        public const string Role = "role";
+        public const string Permissions = "permissions";
+        // Add more default claims as needed
+    }
+}
