+++ CloudServer.Common/Dto/UserDto/UserDto.cs
@@ -1,24 +1,23 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CloudServer.Data.Dto.UserDto
+namespace CloudServer.Common.Dto.UserDto
 {
-    public class UserInputDto
+    public class UserDto
     {
-        [Required]
+        public string Id { get; set; }
         public string Username { get; set; }
-        [Required]
+        public string Email { get; set; }
         public string PhoneNumber { get; set; }
-        [Required]
         public string Address { get; set; }
-
-        public UserInputDto()
+        public UserDto()
         {
+            Id = string.Empty;
             Username = string.Empty;
+            Email = string.Empty;
             PhoneNumber = string.Empty;
             Address = string.Empty;
         }
