+++ CloudServer.Common/Dto/UserDto/UserInputDto.cs
@@ -1,23 +1,24 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace CloudServer.Data.Dto.UserDto
+namespace CloudServer.Common.Dto.UserDto
 {
-    public class UserDto
+    public class UserInputDto
     {
-        public string Id { get; set; }
+        [Required]
         public string Username { get; set; }
-        public string Email { get; set; }
+        [Required]
         public string PhoneNumber { get; set; }
+        [Required]
         public string Address { get; set; }
-        public UserDto()
+
+        public UserInputDto()
         {
-            Id = string.Empty;
             Username = string.Empty;
-            Email = string.Empty;
             PhoneNumber = string.Empty;
             Address = string.Empty;
         }
